requirements.txt
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Flask==2.2.3
fonttools==4.39.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
packaging==23.1
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
threadpoolctl==3.1.0
Werkzeug==2.2.3
